feat: disable sync for dev mode

+++ apps/backend/src/common/entities/entity-syncer/crud-syncer.ts
@@ -13,22 +13,31 @@ export class CRUDSyncer<Entity extends MagexDatabaseEntity>
   implements EntitySubscriberInterface<Entity>
 {
   constructor(@Inject(MagexService) protected magexService: MagexService) {}
+  protected CRUDConfig = {
+    beforeInsert: true,
+    beforeUpdate: true,
+    beforeSoftRemove: true,
+    beforeRecover: true,
+  };
+
   async beforeInsert(event: InsertEvent<Entity>) {
-    console.count('beforeInsert' + event.entity);
+    if (!this.CRUDConfig.beforeInsert) return;
     if (event.entity.lastSyncAt) return;
     await event.entity.createMagexRecord(this.magexService);
   }
 
   async beforeUpdate(event: UpdateEvent<Entity>) {
-    console.count('beforeUpdate');
+    if (!this.CRUDConfig.beforeUpdate) return;
     await event.entity.updateMagexRecord(this.magexService);
   }
 
   async beforeSoftRemove(event: RemoveEvent<Entity>) {
+    if (!this.CRUDConfig.beforeSoftRemove) return;
     await event.entity.deleteMagexRecord(this.magexService);
   }
 
-  async beforeRecovered(event: RecoverEvent<Entity>) {
+  async beforeRecover(event: RecoverEvent<Entity>) {
+    if (!this.CRUDConfig.beforeRecover) return;
     await event.entity.createMagexRecord(this.magexService);
   }
 }
